Working REST API for blog posts with id, title and main content
Made it so addPost.php creates a unique id for the posts and saves it in an array, it increases the id with 1 everytime there is a new post. It also saves the title and the content of the post inside $_SESSION as an array.
Made it so viewPost.php checks for an ID in $_GET and then sends back a JSON object with json_encode, so that the info saved in session can be used later by javascript code.
Script.js added a function with a loop to loop through all posts and show then in corresponding divider, this funciton also parses the JSON from the viewPost.php file. Created a function to show every individual post in their own div with a <h1> tag for title and a <p> tag for content.